+++ src/prelude.rs
@@ -1,9 +1,13 @@
 //! Commonly used types
 
-use crate::compute::LayoutAlgorithm;
-use crate::layout::{SizeAndBaselines, SizingMode};
+#[cfg(feature = "flexbox")]
+use crate::{
+    compute::LayoutAlgorithm,
+    layout::{SizeAndBaselines, SizingMode},
+};
 
 /// Apply the flexbox algorithm and recursively layout the specified node
+#[cfg(feature = "flexbox")]
 #[inline(always)]
 pub fn layout_flexbox(
     tree: &mut impl LayoutTree,
@@ -28,8 +32,8 @@ pub use crate::{
     layout::Layout,
     node::{Node, Taffy},
     style::{
-        AlignContent, AlignItems, AlignSelf, AvailableSpace, Dimension, Display, FlexDirection, FlexWrap,
-        JustifyContent, JustifyItems, JustifySelf, LengthPercentage, LengthPercentageAuto, Position, Style,
+        AlignContent, AlignItems, AlignSelf, AvailableSpace, Dimension, Display, JustifyContent, JustifyItems,
+        JustifySelf, LengthPercentage, LengthPercentageAuto, Position, Style,
     },
     style_helpers::{
         auto, fit_content, max_content, min_content, percent, points, zero, FromFlex, FromPercent, FromPoints,
@@ -37,6 +41,9 @@ pub use crate::{
     },
     tree::LayoutTree,
 };
+
+#[cfg(feature = "flexbox")]
+pub use crate::style::{FlexDirection, FlexWrap};
 
 #[cfg(feature = "grid")]
 pub use crate::style::{