+++ RELEASES.md
@@ -1,5 +1,12 @@
 # Release Notes
 
+## Unreleased
+
+### Changes
+
+- The Flexbox algorithm has now been moved behind the `flexbox` feature. The `flexbox` feature is enabled by default.
+- The `justify_self` property has been moved behind the `grid` feature.
+
 ## 0.3.11
 
 ### Fixes